requirements.txt
asarPy==1.0.1
asgiref==3.8.1
certifi==2026.1.4
cffi==1.17.1
charset-normalizer==3.4.4
cryptography==45.0.2
Cython==3.0.12
Django==5.2.1
django-appconf==1.1.0
django-cryptography==1.1
django-debug-toolbar==5.2.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
idna==3.11
Mako==1.3.10.dev0
Markdown==3.5.2
MarkupSafe==3.0.2
meson==1.5.2
packaging==25.0
pillow==12.1.0
pycairo==1.28.0
pycparser==2.22
Pygments==2.19.1
PyGObject==3.50.1
python-dateutil==2.9.0.post0
python-decouple==3.8
reportlab==4.4.7
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.6.3
wheel==0.46.1
